etl_webapp/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
kombu==5.3.7
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.45
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
